move functions/constants from builder to app_manager.data_source_meta
