+++ corehq/apps/userreports/reports/builder/__init__.py
@@ -23,21 +23,6 @@ FORM_METADATA_PROPERTIES = [
     ('deviceID', 'Text'),
 ]
 
-DATA_SOURCE_TYPE_CASE = 'case'
-DATA_SOURCE_TYPE_FORM = 'form'
-DATA_SOURCE_TYPE_CHOICES = (
-    (DATA_SOURCE_TYPE_CASE, _("Cases")),
-    (DATA_SOURCE_TYPE_FORM, _("Forms")),
-)
-DATA_SOURCE_DOC_TYPE_MAPPING = {
-    DATA_SOURCE_TYPE_CASE: 'CommCareCase',
-    DATA_SOURCE_TYPE_FORM: 'XFormInstance',
-}
-
-
-def get_data_source_doc_type(data_source_type):
-    return DATA_SOURCE_DOC_TYPE_MAPPING[data_source_type]
-
 
 def make_case_data_source_filter(case_type):
     return {
