+++ corehq/apps/userreports/tests/test_report_builder.py
@@ -6,10 +6,10 @@ from django.test import TestCase
 from mock import patch
 
 from corehq.apps.app_manager.models import Application, Module
-from corehq.apps.app_manager.tests.app_factory import AppFactory
+from corehq.apps.userreports.app_manager.data_source_meta import DATA_SOURCE_TYPE_FORM, DATA_SOURCE_TYPE_CASE
 from corehq.apps.userreports.dbaccessors import delete_all_report_configs
 from corehq.apps.userreports.models import DataSourceConfiguration, ReportConfiguration
-from corehq.apps.userreports.reports.builder import DATA_SOURCE_TYPE_FORM, DATA_SOURCE_TYPE_CASE
+
 from corehq.apps.userreports.reports.builder.columns import MultiselectQuestionColumnOption
 from corehq.apps.userreports.reports.builder.forms import (
     ConfigureListReportForm,
